-Replaced FolderBrowserDialog to OpenFileDialog (Deals with deep nesting) -Added drag and drop

+++ Form1.cs
@@ -13,6 +13,7 @@ namespace gmod_audio_converter
         public Form1()
         {
             InitializeComponent();
+            this.AllowDrop = true;
             bool ffmpegExists = File.Exists(@"ffmpeg.exe");
             if (!ffmpegExists) { MessageBox.Show("You are missing one or more dependencies. This application will not function without them.", "Error! Missing Dependencies", MessageBoxButtons.OK, MessageBoxIcon.Exclamation); }
             ffmpegCheck.Checked = ffmpegExists;
@@ -28,10 +29,14 @@ namespace gmod_audio_converter
 
         private void SelectFolderButt_Click(object sender, EventArgs e)
         {
-            FolderBrowserDialog source = new FolderBrowserDialog();
+            OpenFileDialog source = new OpenFileDialog();
+            source.ValidateNames = false;
+            source.CheckFileExists = false;
+            // Always default to Folder Selection.
+            source.FileName = "Folder";
             if (DialogResult.OK == source.ShowDialog())
             {
-                string path = source.SelectedPath;
+                string path = Path.GetDirectoryName(source.FileName);
                 System.Console.WriteLine(path);
                 SourcePath.Text = path;
                 SourcePath.Enabled = true;
@@ -69,7 +74,10 @@ namespace gmod_audio_converter
                         progressBar.Value += 100 / count;
                         ErrorLbl.Text = "(" + iteration + " / " + count + ") Encoding " + Path.GetFileNameWithoutExtension(file);
                         process.WaitForExit();
-                        File.Delete(file);
+
+                        if (shouldDelete.Checked)
+                            File.Delete(file);
+
                         System.Console.WriteLine(process.StartInfo.Arguments);
                         if (iteration >= count)
                         {
@@ -114,5 +122,58 @@ namespace gmod_audio_converter
             Directory.Delete(source, true);
         }
 
+        private void Form1_DragDrop(object sender, DragEventArgs e)
+        {
+            progressBar.Maximum = 100;
+            progressBar.Value = 0;
+
+            string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
+            int count = files.Length;
+            int iteration = 0;
+            string codec;
+
+            foreach (string file in files)
+            {
+                if (!file.EndsWith(".mp3") && !file.EndsWith(".wav")) continue;
+
+                string path = Path.GetDirectoryName(file);
+                Directory.CreateDirectory(path + "\\encode\\");
+
+                var out_path = path + "\\encode\\";
+                var source_path = path;
+                if (Equals(FormalSelect.SelectedItem, ".wav")) { codec = "pcm_s16le"; } else { codec = "libmp3lame"; }
+
+                Process process = new Process();
+                process.StartInfo.FileName = "ffmpeg";
+                process.StartInfo.Arguments = "-y -i \u0022" + file + "\u0022 -c:a " + codec + " -ar " + SampleSelect.SelectedItem + " \u0022" + out_path + Path.GetFileNameWithoutExtension(file) + FormalSelect.SelectedItem + "\u0022";
+                process.StartInfo.UseShellExecute = false;
+                process.StartInfo.CreateNoWindow = true;
+                process.Start();
+                iteration++;
+                progressBar.Value += 100 / count;
+                ErrorLbl.Text = "(" + iteration + " / " + count + ") Encoding " + Path.GetFileNameWithoutExtension(file);
+                process.WaitForExit();
+                
+                if (shouldDelete.Checked)
+                    File.Delete(file);
+
+                System.Console.WriteLine(process.StartInfo.Arguments);
+                if (iteration >= count)
+                {
+                    ErrorLbl.Text = "Encoding complete. Converted " + iteration + " files to " + FormalSelect.SelectedItem + " at " + SampleSelect.SelectedItem + " Hz";
+                    progressBar.Value = 0;
+
+                    if (Directory.Exists(out_path))
+                    {
+                        MoveDirectory(out_path, source_path);
+                    }
+                }
+            };
+        }
+
+        private void Form1_DragEnter(object sender, DragEventArgs e)
+        {
+            e.Effect = DragDropEffects.Copy;
+        }
     }
 }
